Date: July 31, 2003 Modification of the method GiveInfo.
In the old version, the test for updating index in the reverse cases when the vector is not in the same order than the forward one (sensors NStart, NG2, SPred) have to be size-1 instead of size. In the same cases, the evaluation of lower_bound have to be tested to correctely treat cases before the first site and after the last one.
Beware: lower_bound used with the greater function, for the position 0 returns 0 even if 0 is not in the vector, then *iter == pos is not a sufficient condition for the presence of a site.
The method has changed. Now, if GiveInfo if called with a position on the sequence that differs from  the position given in the last called plus 1, then the indexes are updated.
Inconvenient: in case of small difference between 2 calls positions, index are updated even if not necessary.

Some reference files for tests are changed.