Adding the ability to automagically fill ontology term if a étrsbcript" structure is defined: all CDS having a parent transcript features are consecutive exons that can be categorized as First/Internal/Last or Sngl.
The code is beta, it assumes that all the children of a feature are sorted according to position. If the parameter below is set to "", the code will not do anything.

A new parameter AnnotaStruct.TranscriptFeature has been introduced. It must be setto the name of the feature that identifies "transcript" units (mRNA for Eugene and GeneId, transcript for augustus)